+++ src/index.html
@@ -154,6 +154,6 @@
             </a>
         </div>
     </footer>
-    <script src="js/index.js" type="module"></script>
+    <script src="index.js" type="module"></script>
 </body>
 </html>
