+++ src/public/components/BannerComponent.js
