+++ src/public/components/CarouselContent.js
