+++ src/public/components/IssueComponent.js
