+++ src/public/components/LoadingComponent.js
