+++ src/public/components/NaviRemoteComponent.js
