+++ src/public/events/footerEvent.js
@@ -1,12 +1,12 @@
-import { $ } from "./utils.js";
+import { $ } from "../utils/utils.js";
 
-export const footer = {    
+export const footerEvent = {
     detailSlide() {
         const businessInfoEl = $(".footer__business_info");
         const businessDetails = $(".footer__details");
         businessInfoEl.addEventListener("click", ()=> {
-            businessInfoEl.classList.toggle("footer__business_info_off")
-            businessDetails.classList.toggle("no_display")
+            businessInfoEl.classList.toggle("footer__business_info_off");
+            businessDetails.classList.toggle("no_display");
 
             businessDetails.classList.contains("no_display") ? 
             businessInfoEl.style.setProperty("--background", "url(../images/icon_trace_off.png)") :
