+++ src/public/utils/utils.js
@@ -1,14 +1,7 @@
 export const $ = (className) => document.querySelector(className);
 export const days = ['월', '화', '수', '목', '금', '토', '일', '완결'];
 export const today = days[new Date().getDay() - 1];
-import { serverURL } from "./constants.js";
-
-export const getWebtoonData = (url) => {
-    // fetch test func
-    fetch(url)
-        .then(response => response.json())
-        .then(data => console.log(data));
-}
+const serverURL = 'http://localhost:3000';
 
 export const fetchData = (path, id) => {
     const url = id !== undefined ? `${serverURL}/${path}/${id}` : `${serverURL}/${path}`;
@@ -39,3 +32,5 @@ export const scrollMove = (moveY) => {
         }
     }, 10);
 }
+
+export const delay = (time) => new Promise(resolve  => setTimeout(resolve, time * 1000));
