+++ src/public/views/handleCategoryMenu.js
@@ -1,13 +1,13 @@
-import { $, fetchData, days, today } from "./utils.js";
+import { $, fetchData, days, today } from "../utils/utils.js";
 import { ContentListInfos } from "./ContentListInfos.js";
-import { categoryObj } from "./constants.js";
-import { NaviRemoteComponent } from "../component/NaviRemoteComponent.js";
-import { nav } from "./nav.js";
+import { categoryObj } from "../constants/constants.js";
+import { NaviRemoteComponent } from "../components/NaviRemoteComponent.js";
+import { navEvent } from "../events/navEvent.js";
 
 const commonGenreRender = async (element, categoryId) => {
     const mainInner = $('.main__inner');
     mainInner.insertAdjacentHTML("beforeend", NaviRemoteComponent());
-    nav.remoteClick();
+    navEvent.remoteClick();
 
     const issueData = await fetchData('category/issue', categoryId);
     const recommendData = await fetchData('webtoon/recommend', categoryId);
